Item get api 개발

+++ src/main/kotlin/com/example/shoppingapi/domain/CartItem.kt
@@ -4,5 +4,5 @@ import com.example.shoppingapi.domain.document.Item
 
 data class CartItem(
     val item: Item,
-    val quantity: Int = 1,
+    val quantity: Double = 1.0,
 )
